requirements.txt
torch==2.3.1
torchaudio==2.3.1
soundfile == 0.12.1
faster_whisper==1.0.1
pytorch_lightning==2.2.1
h5py==3.11.0
numpy==2.0.0
pandas==2.2.2
PyYAML==6.0.1
tqdm==4.66.4
wandb==0.17.4
